+++ hydromt/data_adapter/dataframe.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import os
 from os.path import join
 import numpy as np
 import pandas as pd
@@ -148,8 +149,18 @@ class DataFrameAdapter(DataAdapter):
         For a detailed description see: :py:func:`~hydromt.data_catalog.DataCatalog.get_dataframe`
         """
 
+        # Extract storage_options from kwargs to instantiate fsspec object correctly
+        if "storage_options" in self.kwargs:
+            kwargs = self.kwargs["storage_options"]
+            # For s3, anonymous connection still requires --no-sign-request profile to read the data
+            # setting environment variable works
+            if "anon" in kwargs:
+                os.environ["AWS_NO_SIGN_REQUEST"] = "YES"
+            else:
+                os.environ["AWS_NO_SIGN_REQUEST"] = "NO"
+        _ = self.resolve_paths(**kwargs)  # throw nice error if data not found
+
         kwargs = self.kwargs.copy()
-        _ = self.resolve_paths()  # throw nice error if data not found
 
         # read and clip
         logger.info(f"DataFrame: Read {self.driver} data.")